Function ds. Deprecated.

Deprecated. Use function downsize instead. See help_downsize for help.

ds(
  big,
  small = NULL,
  downsize = getOption("downsize"),
  warn = TRUE,
  random = FALSE,
  length = NULL,
  dim = NULL,
  ncol = NULL,
  nrow = NULL
)

Arguments

big

Object to return if downsize is FALSE or NULL.

small

Object to return if downsize is TRUE and all subsetter arguments such as length and dim are NULL.

downsize

TRUE/FALSE value (NULL counts as FALSE), whether to replace big with a downsized object. Defaults to the global option downsize, which you can check with getOption("downsize") or the scaling function and set with functions downsize, scale_down or scale_up.

warn

TRUE/FALSE option to warn the user if big and small are identical or big is smaller in memory than small.

random

If TRUE, take a random subset of big instead of the first few elements. For example, if nrow == 3, take a random 3 rows instead of the first 3.

length

Downsize big to this length if downsize is TRUE.

dim

Downsize big to these dimensions if downsize is TRUE.

ncol

Downsize big to this number of columns if downsize is TRUE.

nrow

Downsize big to this number of rows if downsize is TRUE.

Value

A downsized object if downsize is TRUE and big otherwise.

See also